vasprintf_compat.h: spin-off this compat header ; use math_compat.h as template
