+++ Makefile.am
@@ -33,6 +33,7 @@ libjson_cinclude_HEADERS = \
 	linkhash.h \
 	math_compat.h \
 	strdup_compat.h \
+	vasprintf_compat.h \
 	printbuf.h \
 	random_seed.h
