+++ printbuf.c
@@ -27,6 +27,7 @@
 
 #include "debug.h"
 #include "printbuf.h"
+#include "vasprintf_compat.h"
 
 static int printbuf_extend(struct printbuf *p, int min_size);
 
@@ -109,47 +110,6 @@ int printbuf_memset(struct printbuf *pb, int offset, int charvalue, int len)
 
 	return 0;
 }
-
-#if !defined(HAVE_VSNPRINTF) && defined(_MSC_VER)
-# define vsnprintf _vsnprintf
-#elif !defined(HAVE_VSNPRINTF) /* !HAVE_VSNPRINTF */
-# error Need vsnprintf!
-#endif /* !HAVE_VSNPRINTF && defined(WIN32) */
-
-#if !defined(HAVE_VASPRINTF)
-/* CAW: compliant version of vasprintf */
-static int vasprintf(char **buf, const char *fmt, va_list ap)
-{
-#ifndef WIN32
-	static char _T_emptybuffer = '\0';
-#endif /* !defined(WIN32) */
-	int chars;
-	char *b;
-
-	if(!buf) { return -1; }
-
-#ifdef WIN32
-	chars = _vscprintf(fmt, ap)+1;
-#else /* !defined(WIN32) */
-	/* CAW: RAWR! We have to hope to god here that vsnprintf doesn't overwrite
-	   our buffer like on some 64bit sun systems.... but hey, its time to move on */
-	chars = vsnprintf(&_T_emptybuffer, 0, fmt, ap)+1;
-	if(chars < 0) { chars *= -1; } /* CAW: old glibc versions have this problem */
-#endif /* defined(WIN32) */
-
-	b = (char*)malloc(sizeof(char)*chars);
-	if(!b) { return -1; }
-
-	if((chars = vsprintf(b, fmt, ap)) < 0)
-	{
-		free(b);
-	} else {
-		*buf = b;
-	}
-
-	return chars;
-}
-#endif /* !HAVE_VASPRINTF */
 
 int sprintbuf(struct printbuf *p, const char *msg, ...)
 {
